requirements.txt
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msal==1.31.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.1.3
